requirements.txt
# Core Streamlit app
streamlit>=1.28.0

# Audio processing
librosa>=0.10.0
soundfile>=0.12.0
praat-parselmouth>=0.4.3
torch>=2.0.0
openai-whisper>=20231117

# Data & ML
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
xgboost>=1.7.0
joblib>=1.3.0
scipy>=1.10.0

# Visualization
plotly>=5.15.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Streamlit audio recorder (correct package name)
audio-recorder-streamlit>=0.0.8
